move grid_sample code from modules to functional
